feat(ui): show the plan tier where Pacer shows accounts
The tier reached the API and the skill and stopped there — the app's own
surfaces still showed the coarse family. Both places that display an account
pass a `plan` string into the shared `AccountRow`, so both now pass the tier-
aware label: Settings → Tokens → Accounts, and the dashboard's per-account
group on the Rate-limit pace card, which reads `Max 20×` where it read `max`.

`PlanLabel` owns the parsing. It was a computed property on `OAuthCredential`,
which meant the API builder had to construct a throwaway credential to render a
stored account — a tell that the logic was on the wrong type. Three surfaces
need the same answer and a plan rendered one way in Settings and another on the
dashboard is the drift that starts with a copied helper.

Screenshots regenerated; the fixture now carries a tier so the gallery shows
what a real install shows. Most of the churn in those PNGs is the harness's
time-varying data ("read 30s ago"), not this change.

## Why the second account still reads "Max"

Traced with a temporary probe rather than guessed at, and it is not a bug in
this path. Pacer never polls that account: it ingests the switcher's usage
cache instead, deliberately, so two clients do not poll the same token and
rate-limit each other. That cache carries percentages and nothing about the
account behind them — verified, it has no plan field at all — so there is no
credential in hand and nothing to read a tier from. The probe confirms it:
every polling lane resolves to the *other* org.

A plan is never overwritten with nil, so the detail arrives and stays the first
time that account is the live login. The switcher-cache call site now says all
of this where the next reader will be standing.

960 tests pass.
