+++ PacerCore/Sources/PacerCore/API/PacerAccounts.swift
@@ -292,9 +292,7 @@ public enum PacerAccountsBuilder {
                     organizationName: account.organizationName,
                     subscriptionType: account.subscriptionType,
                     rateLimitTier: account.rateLimitTier,
-                    plan: OAuthCredential(accessToken: "", expiresAt: nil,
-                                          subscriptionType: account.subscriptionType,
-                                          rateLimitTier: account.rateLimitTier).planLabel,
+                    plan: account.planLabel,
                     isActive: account.isActive,
                     unattributed: false,
                     firstSeenAt: account.firstSeenAt,
